Added code for a memory bus module
This was done in order to decouple the system components from one
another and make testing easier to do.

The memory bus will also encapsulate the memory-mapped IO and
reroute memory read/writes from the CPU to the appropriate devices
without the cpu even being aware that the devices exist.