Added support for memory bus use in the memory module.
Removed FIXME's for memory-mapped device IO in the memory module (this
        will now be encapsulated by the memory bus).

Created new memory_cycle() function that emulates what happens in RAM
when a single clock cycle elapses. It will return the data the processor
requests after the encoded delay time (which can be adjusted).

Exported a function for printing memory contents to other modules.

Removed #include of memory.h from memory bus (it shouldn't depend on the
        devices attached to it).