Removed memory module dependency in cpu module.
Added support for memory bus module use in cpu module. The cpu no longer
needs to have knowledge of devices it communicates with. Programs can
now write to memory-mapped IO locations, and the memory bus will handle
the rest.

Added function to detect when an instruction has finished executing in
the cpu module and exported it for use in the computer. This is to make
single-stepping easier.

Created a finite state machine for the different instruction execution
phases. The CPU FSM gets processed every clock cycle and now more
accurately reflects what happens when things like delays fetching data
from RAM occur. This new cpu implementation is not pipelined.