Added debug helper functions in computer module
Added support for memory bus use by the computer module.

Added "factory"/"build" functions for the computer and cpu modules. The
reason for this is so that the "constructors" of our computer and cpu
modules no longer do real work. The constructors now set up the
dependencies of each module by having them passed in via dependency
injection. I've never used this technique before, but I am trying it out
here to see if the claims in the research I have done are true: namely
that dependency injection makes testing modules/classes easier since the
constructor does no real work (which is good because typically
        constructors can't be overridden).

The computer module now depends on all of the subcomponents such as the
memory bus, RAM, and the CPU (and later other devices like timers and
        graphics). This is as it should be since real computers need all
of these things to work. Further, having the computer module depend on
all these other submodules and control simulation at the clock cycle
level helps get rid of weird inter-module dependencies like I had before
(such as the CPU depending on the memory module). This makes for a
cleaner and more flexible design.