Removed hardware submodule dependencies in main.c
NOTE: This is now a working revision of the simulator.

The main program now only needs to know about a "computer" which takes
care of all of the hardware dependencies on its own. All main needs to
care about currently is creating a computer, loading a program on it,
running the program/single-stepping, and outputting any useful debug
information. In the future, when graphics support is added, all SDL
setup and GUI related code will reside here.