Added a "preprocessor assembler" to the project to facilitate testing the CPU. Previously, test programs would be directly encoded in an array in the main program. Manually calculating the instruction encodings is tedious and error-prone, so I added this "assembler" as a shortcut for testing purposes. The advantage it has is that it's much easier to see the individual instructions, but it can't deal with labels, so those will still need to be calculated by hand.
The "assembler" is used by using function-like macros (that are defined
in "preprocessor_assembler.h") in place of the manually-encoded integer
constants of a uint32_t array that represents our program. The macros
will expand to the correct values when the preprocessor is run.

I will eventually add a real assembler for this machine, but for quickly
testing small sequences of instructions, this is much better than
before.