THE CODE IS IS A NON-FUNCTIONAL STATE
NOTE: At some point I need to draw out a memory map for this system and
its peripherals.

I wanted to break movement of the SDL code out of main into discrete
steps, so that's why this is a non-functioning commit.

The SDL graphics subsystem has been moved from main.c into its own
graphics module. The computer module now has awareness of the graphics
system and there is some debug-only code in place to make sure that we
can actually still spawn the window and write the frame-buffer contents.

The graphics subsystem still has no knowledge of the memory bus and vice
versa. This will need to be fixed. Eventually we will want this system
to handle double buffering for our computer in a transparent way as
well.

Right now, we can spawn a window and dump framebuffer contents to the
window, but the IO is still commented out in main, so we have no way to
gracefully terminate the program. I will move that code into an IO
module next, and then after that, I will figure out how to interface the
two systems to the memory bus.