The system works again, with some hacks.
I just wanted to get the program quickly back into a state where I could
gracefully terminate without needing to resort to the OS killing the
process. We can exit the program again now by clicking the "x" in the
window or typing "q" on the keyboard. This later part will need to be
changed to some other key because eventually the keyboard subsystem will
need to pass incoming keycodes to the processor being simulated.

So, in summary, I moved the SDL code to handle keyboard events into a
new keyboard module. There is a dirty hack in the code so that it can
clear the "simulation_running" flag. This will get fixed shortly.
Additionally, there is some hacky code that I put in the computer module
in order to test that the new graphics and keyboard subsystems are
working. I will eventually need to figure out when is the correct time
to render the frames and poll the keyboard. Maybe they can both happen
at the same time since there is no way that a human can type faster than
60 keys/second, and I want to have a steady framerate of 60Hz.