Changed the way SDL was being linked and included in the code.
This is supposed to make things more "portable" according to what I'd
read about SDL online. Now, if someone else wants to compile this
project on a different machine, theoretically they just need to change
the SDL include folder in the makefile to make things work.

Added a memory map header file to the project.

The memory map header outlines how devices are mapped into the physical
memory space of the computer. I used this information to update the
memory_bus module so that it can properly route information according to
the value located on the memory address lines.

NOTE: The keyboard and graphics modules currently interact with the
program but aren't technically "installed correctly" in our computer at
this point. What this means is that although we are currently receiving
keyboard input and rendering to the window, the two modules are not
interacting with the rest of the computer via the bus yet. This isn't a
problem at the moment though because the CPU doesn't have support for
load and store instructions yet, which are necessary for the CPU to
interact with these memory-mapped peripherals. I'll go back and fix this
after I've implemented some simple load/store instructions.