Big commit: All changes are related to adding support for LOAD and LOADR instructions to the CPU.
Unfortunately all of these changes to various files had to happen more
or less together to ensure that these new instructions work. At this
point, I have successfully tested the LOAD instruction with a short
program, but LOADR is not yet tested.

I shortened the memory access delay time in memory.c for debugging
purposes, as I was printing out the cycle names as they executed. I'll
change this back later.

I updated the cpu struct so that it can keep track of information
associated with the new memory instruction formats (i.e. the base
register, and pc-relative and base-register-relative offsets). I added
two utility functions to the cpu_op module so that the cpu module
wouldn't need to know the exact encoding of instructions to determine if
they are memory instructions. This makes rearranging the opcodes easier
if I decide to do that later.

By far the biggest change that had to occur was modifying the CPU FSM
code. I had to add two new stages called "MEMORY1" and "MEMORY2". As the
name implies, they deal with memory accesses and are only used by
instructions that read/write to memory. Only load instructions have been
encoded so far, so further tinkering with the FSM code may be needed for
the store instructions.

The reason there are two memory stages is the same as why there are two
fetch stages, and I feel its largely due to a weakness in how this
program is written. Basically, since the code separates the bus, the cpu
and the memory systems into separate modules with minimal dependecies,
so the function that implements a fetch/memory stage must write its
values to the MAR and execute to completion before the bus_cycle
function can execute and route values to memory/peripherals, which means
that the requested data is not available back in the fetch/memory
function at the time it was executing, and therefore the cpu isn't
finished fetching/accessing memory and is not yet ready to move to the
decode/execution stage, which implies that an additional fetch/memory
stage must be put in place to finish processing the data.

Ultimately, even though this makes the cpu FSM conceptually less clean,
it allows me to minimize inter-module dependencies, which is really
important. Also, in a sense doing things this way (as opposed to just
having the cpu access an array representing RAM at the moment it's
needed) makes for a more realistic simulation. This is because now we
can simulate arbitrary memory access delay times easily since the second
memory/fetch stages can just transition to themselves while they wait
for the data access to finish. Furthermore, it may also make it easier
for me to add caching to the simulation later. Finally, looking at the
LC3b FSM in Yale Patt's book shows a cpu that has at least 2 fetch
stages / memory access stages, so I don't feel quite so weird about
having more than one of each type of stage myself.