VICTORY! The cpu can now execute the test program I wrote in main and makes a 6x6 blue square on the display window!
It doesn't seem like much, but that little square means that the
graphics peripheral has been memory mapped properly, the cpu has support
for loading addresses and writing to them, and can do basic bit-wise
manipulations. The next steps from here will be to add support for jumps
and branches, and to connect other peripherals like the keyboard and
maybe some timer modules. Following that, I will need to add support for
interrupts (both hardware and software), so that I can do more
interesting things with this machine.

As for what changed in the code since the last commit...

I installed the graphics module in the computer module (i.e. added the
graphics_cycle() call along with the bus_cycle() and cpu_cycle() calls).

I fixed an annoying warning in the computer module (it was about the
wrong printf format specifier).

I removed some debug code in the cpu module and fixed 2 bugs:
    1) The first bug was that I was retrieving the base register for
    instructions from the wrong bit-positions in the instruction, which
    caused the wrong registers to be read from during address calculation
    2) The second bug was a segfault that was being caused by improperly
    adding a pointer value (that should have been dereferenced) to an
    offset value, which caused the program to crash

In the graphics module, I removed some debug function calls, removed a
call to graphics_update() that i'd forgotten about that was causing the
program to crash (since the semantics of the function had changed and it
was no longer being used properly). I changed graphics_update() so that
it is aware of its placement in the memory map and adjusts the address
being written to by the correct offset. Finally, I commented out an
unused function to make gcc shut up about it not being used.

In the main module, I wrote a simple program for the CPU to write a blue
square to the output, and I adjusted the rest of the main() function to
increase the amount of debug information gleaned from the program's
execution.