Added support for pc-relative jump instructions to the cpu and the preprocessor_assembler (and its unit tests).
Added 2 purely mnemonic instructions CLEAR and HCF to facilitate
clearing a register and halting the program by spinning forever,
respectively. These two instructions are just aliases for special XOR
and JUMP instructions, they don't have their own opcodes.

Added a generic sign-extension function because I was about to write a
3rd specific sign-extension function, and doing this is tedious and
error prone. The code in cpu.c is comming close to needing some
refactoring to get rid of the magic numbers, but I will hold off on that
for now. I also changed the sign-extension of the offset bits to happen
in the decode phase of the instruction, so the internal processor
registers that hold these offsets will now hold the sign-extended values
instead of raw values taken from the instruction.

I started to write a little program to clear the frame buffer to white,
but I realized that to do this easily, I need to add support for the add
instruction first, at a bare minimum. I also need to add support for
some sort of branch instructions soon to do anything useful.