Wrote a program to make a blue square on the screen using branch instructions. It runs SLOWLY. I will need to think about what I will need to do to speed up performance of this simulator, but at the moment I don't know how to use linux program profiling tools. So, that will be one of the next things to look into.
NOTE: The little program I wrote seems to work (aside from running
slowly), but it always seems to draw the square in the upper left
corner of the screen, and I'm not sure why or how to fix that yet.

I also removed a bunch of debug printf statements to get rid of most of
the extraneous terminal output.