Added support for the xor instruction to the cpu.
Added a simple branch back to the end of the beggining of the "blue
square" program. With XOR working correctly now, the CLEAR instructions
at the start of the program should actually have an effect, and the
program will overwrite the same spot in memory again and again instead
of crashing like it did before I added XOR support. To be clear, since
there was no XOR support, the registers weren't actually cleared, and
things worked the first time around because the registers are
initialized to 0x00 by the simulator at start up, but weren't clearing
on subsequent iterations, and eventually the program would write beyond
the frame buffer to a _currently_ invalid memory location (because there
was no memory installed there yet) and would crash.