Added internal register to track the status information of the currently executing process (including interrupt processing status information).
Added code to keep a new interrupt from being launched (or even the same
interrupt from launching twice) to prevent bugs. For the moment, nested
interrupts are not supported on this system.

Refactored the code that handles setting the condition code bits so that
it now uses the bit-twiddling macros instead of raw operations.

Added code to check if an interrupt is currently in flight and to set
the interrupt processing status.