I changed the semantics for how the interrupt controller is used. The interrupt controller now uses a queue internally to process/sequence all interrupt requests, which simplifies the interrupt controller code somewhat. Queues can only be accessed from their head, which means it is no longer possible for hardware devices to clear their interrupts (since they may be somewhere in the middle of the queue). The ISR code to handle the various interrupts will still need to clear the appropriate bits to prevent another interrupt from immediately being requested by the peripheral, but essentially, an interrupt is "cleared" when it is "gotten" from the queue by the CPU.
I removed code from timer.c to clear the timer interrupt request (since
that is no longer possible). I also changed the timer code so that it
will now only request an interrupt on the same cycle that it overflows
instead of requesting an interrupt asynchronously to the overflow. It
used to be that as long as the timer interrupt flag was asserted and
timer interrupts were enabled, the timer would make an interrupt
request. This didn't hurt anything under the old interrupt controller
semantics, but under the new system, it will cause the queue to
overflow.

We will see if this new change to the interrupt controller proves
problematic or not...