I refactored the "AND" tests in tests/cpu_instruction_tests.cpp to use test_single_instruction(). Unfortunately not much is saved by way of lines of code because just feeding a bunch of constant literals to the function as parameters would be really hard to read.
I wanted to use "designated initializers" and "compound literals" to
simulate named function parameters, but I can't do that because C++
doesn't support this (and the tests are technically written in C++).

I thought about just wrapping all of these parameters into a simple
struct, initializing the struct fields one at a time, and passing that
struct to test_single_instruction() as a single parameter bundle, but
when I tried doing that, for some strange reason it caused my tests to
end up in an infinite loop. I tried to track down why this is, but
ultimately decided it was a waste of time and moved forward with the
current setup as seen in this commit.