Added preliminary (i.e. untested) support for the CALL and CALLR instructions. The CALL instruction is a pc-relative addressed instruction that implements calling a subroutine by jumping to the pc-relative address and saving the return address in R30. CALLR does the same as CALL except the jump target is calculated by using a base register + offset instead of using pc-relative addressing.
I will implement the rest of the CALLR instruction as soon as there is
support for a JUMPR instruction (which will jump with the jump target
being calculated using a base register + offset). Having JUMPR
implemented will also allow me to easily implement the RETURN
instruction, which is merely an alias to "JUMPR R30".